requirements.txt
pillow
customtkinter
openpyxl
screeninfo
pygetwindow
selenium
beautifulsoup4
box-sdk-gen
flask
